chore: move downloadFile util test
